requirements.txt
robotframework
robotframework-requests
robotframework-jsonlibrary
robotframework-faker
psycopg[binary]
psycopg[pool]
